requirements.txt
openai
langchain
langchain-openai
langchain-community
langgraph
llama-index
faiss-cpu
sqlalchemy
pandas
sqlglot
tiktoken
python-dotenv
sentence-transformers
ollama
langchain-anthropic
anthropic